Disable broken code added in the last commit
If you weren't already in a git repository, you got the usual error
message from git, telling you as much, which of course isn't conducive
to the desired operation.

The next step will likely be to, in order:

  1. Check current directory for `.git`. If found, take note of the
     directory in which it was found, then bail.

  2. Iterate over each upper-level directory until `.git` is found, and
     then handle as in the above.

Unless there's a way to shut git up, this will probably have to be it.
Luckily, it might prove to be a more efficient use of resources.