Add small new project called `dayborn`
Pretty nifty, but quite specific, so probably would really only be used
on my programs, since I always use the same header, unless I have to
change the comment character, such as in Vi-IMproved.

I may add this to the TFL module, or have it as a little helper tool
for programs I have, but I'm not entirely sure why, yet.

In any case, it has educational purposes, so that's something.