Correct version strings in WCDL header
Not corrected so much as held back a lot, because I accidentally
grabbed those current versions from my Ubuntu 18.04 laptop.

In a lot of cases, the versions I specify as dependencies may be wildly
unnecessary, but a safety measure to ensure things work and things work
as intended. This can likely be overridden, however.