Add `CPUTemp`, `CoreTempCrit`, & `TempScale` keys
Also updated the database and the man page.

Limitedly, this is just one method by which to determine the CPU core
temperatures, so it's likely that a lot of people will not be able to
use it; if you're such a person, and would like this feature to work
for you, please let me know.