Add `CMOS`, `RAM`, & `RAMMBMin` keys, & add opts
Have also updated the database and man page to reflect the new changes.
I've included some extra information in the man page, too. The key 1118
(from the `CMOS` key) has plenty of information in the database.

Added `--list-keys` and `--list-dbalerts` options. The latter of which
is to be used for the Bash completion script, to complete alerts when
using the `--query|-Q` flag.