Add `--key|-k KEY=VALUE` functionality (yay!)
This has been something I've wanted to see for a while; it'll make
testing and diagnosing problems much easier, as it will while working
on the program.

Of course, for those of us wanting to just quickly override a setting
without going into the `config` file every time, this is perfect.

This option can be provided multiple times, which is even better.