Disable REGEX for `UnwantedPKGs` (addresses bug)
Shame to have to do this, but using `$` in REGEX seemed to be entirely
non-effective, so until I can rectify this, I think it makes more sense
to simply allow the functionality to work without REGEX, at least for
now.

I'll be committing the appropriate changes to the `config` next.