Add functionality to UbuChk to read database-rk
Looks to be working correctly. Had to do make some changes to the
'config' file and how potential infections are ignored; I believe it
was wonky to begin with anyway, so it's more of a fix. Now the user
just needs to enter the file they wish to be ignored by UbuChk.