Add menu to GPreviewer & allow database updating
Now we can update the database without resorting to the terminal. But
do keep in mind that it won't tell you if it succeeded or not. If you
notice a momentary freeze, then it's probably working; shouldn't take
long.

Eventually, I'll include a message window on success, and have it
display an error dialog if it fails. I might even add extra
functionality not in Previewer which shows how much it updated by, or
better yet, how many new reviews were added! That'd be sweet.