Rename project to LAM & add bit more to it
Not much added. Mainly just pseudo-fixed the padding of a widget in
the Settings window's top-most HBox.

The reason I changed the name is because I realised the previous name
was not only too long and generic, but also incorrect, since it's not
'instant'-messaging. It's intended as more of a pseudo-IM experience, -
for local-only use.