Add new Depends project
Partly for reference, partly for educational purposes, and partly due
to border. Depends just works like `command` and `type`, without the
weirdness, such as needing to do all that `>/dev/null 2>&1` guff.

If anyone finds this useful, I might add to it. I can think of a few
things to make it a bit more robust and useful to the user. Never
underestimate the value of dependency-checking.