Use FErr() instead of Err() in UbuChk
Found old use of Err().

Although I plan to go with the boring die() and warn() subs, in hopes
of eventually phasing out the need for TFL.pm in at least UbuChk; this
would result in much more portability, and easier installations.

The downside is that UbuChk would be bigger, but it's not exactly a
huge deal -- not even dealing with MiBs, yet.