Update UbuChk with omission of TFL module
I've also tweaked the logic of a lot of code, preferring CONDITION
and/or ACTION as opposed to the other way around I previously had.

I've had to tweak (and in some cases dumb-ify) some die() calls due to
the way in which they work.

UbuChk used the TFL module quite a bit, so to stop that, I had to bring
a few subroutines into UbuChk itself. This will add a bit of size, but
nothing major. The line count is, I think, around 2.2k.

UbuChk will no longer suggest Cito executions upon missing file(s)
detection; this is due to the way in which die() works, and I couldn't
get it to look 'just right' (thank you, OCD). It will probably make a
return in the future, as it was kinda neat for those using Cito.

As this was quite an impactful change to UbuChk, it's possible this
commit has broken something. It's better off with testing for a while
before it can be built on in any great capacity.