Remove default config & instead use __DATA__
I've also added a 'Verbose' line for when this file is detected in this
version and onwards, just to be kind to the user, but ideally the
installer would take care of this.

I've also made it so that the config header is dynamic, so I needn't
keep updating that, which means fewer chances for blunders.