Update man page & allow searching all arch-types
Before this commit, you had to give the full architecture of a package,
such as 'firefox:amd64', but this is no longer required, as all found
architectures for the given package will now be shown, unless otherwise
specified.

Information in the man page has now been added to reflect this change.
I've also added the AUTHOR section, and fixed the botched formatting of
the OPTIONS section.