Add new DTP project ('drivetemp' Perl)
Quite pleased with this project. It makes use of an awesome feature of
the modern Linux kernel, called _drivetemp_, which itself allows access
to current drive temperatures via the _hwmon_ class.

You can filter by one or more block device names, and has a neat (but
simple) color feature for the temperature, which changes based on the
temperature of the drive; not altogether useful, but kinda pretty.