Add sizes, 1024/1000 switching, & unit switching
Had to change some OPTs around, to allow for `-H` as the short-form of
`--si`, in keeping with other programs, like df(1).

Sizes are determined by the sectors and hardware sector size available
in SYSFS. No human-readable format available yet, but might add that,
eventually.

The header updates 'SIZE' to SIZE + _?, where '?' is the unit
character, such as 'M'. By default it's just SIZE, since we'd just be
showing bytes. The SIZE column is also dynamically sized.

The new `--size=` key=value OPT allows the user to change the size
displayed to a given unit. If using, for example, 'P' (Petabytes) on a
drive so small (IE: 60G) the number becomes unreasonable to output, so
'0.00' will instead be shown.