Make 'boring' default & improve summary
This program needs reviewing. I plan to do this better, as it really
needs improving. The main issue is the over-complication due to
printing and parsing at the same time. I'd like to rework some of it so
that it first gathers the data, _then_ displays the results; this has
the added benefit of making counts less convoluted.